check to see if we are dealing with an extracted aab file when processing unity android aab files

+++ pkg/upload/android-aab.go
@@ -41,28 +41,32 @@ func ProcessAndroidAab(
 	var err error
 
 	for _, path := range paths {
-
 		// Look for AAB file if the upload command was run somewhere within the project root
 		// based on an expected path of ${dir}/build/outputs/bundle/release/${dir}-release.aab
 		// or ${dir}/build/outputs/bundle/release/${dir}-release-dexguard.aab
 		if utils.IsDir(path) {
-			arr := []string{"*", "build", "outputs", "bundle", "release", "*-release*.aab"}
-			path, err = android.FindAabPath(arr, path)
+			if utils.FileExists(filepath.Join(path, "BUNDLE-METADATA")) {
+				aabDir = path
+			} else {
+				arr := []string{"*", "build", "outputs", "bundle", "release", "*-release*.aab"}
+				path, err = android.FindAabPath(arr, path)
 
-			if err != nil {
-				return err
+				if err != nil {
+					return err
+				}
+
+				if filepath.Ext(path) != ".aab" {
+					return fmt.Errorf("%s is not an AAB file/directory", path)
+				} else {
+					aabDir, err = utils.ExtractFile(path, "aab")
+
+					defer os.RemoveAll(aabDir)
+
+					if err != nil {
+						return err
+					}
+				}
 			}
-		}
-		if filepath.Ext(path) != ".aab" {
-			return fmt.Errorf("%s is not an AAB file/directory", path)
-		}
-		logger.Debug(fmt.Sprintf("Extracting %s into a temporary directory", filepath.Base(path)))
-		aabDir, err = utils.ExtractFile(path, "aab")
-
-		defer os.RemoveAll(aabDir)
-
-		if err != nil {
-			return err
 		}
 	}
